refactor(connectivity-forms): split `ConnectivityForm` into reusable components `VoltageLevelConnectivityForm` and `PositionForm` for better modularity and clarity
